pyproject.toml
[tool.poetry]
name = "regioes-vs-estados-vs-cidades"
version = "0.1.0"
description = ""
authors = ["Wesley Borges <djwesleyborges@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.4"
django-ninja = "^1.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
